Add check for feature enable/disable

+++ app/bundles/CoreBundle/Command/AnonymizeIpCommand.php
@@ -13,6 +13,8 @@ namespace Mautic\CoreBundle\Command;
 
 use Doctrine\DBAL\DBALException;
 use Mautic\CoreBundle\Entity\IpAddressRepository;
+use Mautic\CoreBundle\Helper\CoreParametersHelper;
+use Mautic\CoreBundle\Helper\ExitCode;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -29,9 +31,15 @@ class AnonymizeIpCommand extends Command
      */
     private $ipAddressRepository;
 
-    public function __construct(IpAddressRepository $ipAddressRepository)
+    /**
+     * @var CoreParametersHelper
+     */
+    private $coreParametersHelper;
+
+    public function __construct(IpAddressRepository $ipAddressRepository, CoreParametersHelper $coreParametersHelper)
     {
-        $this->ipAddressRepository = $ipAddressRepository;
+        $this->ipAddressRepository  = $ipAddressRepository;
+        $this->coreParametersHelper = $coreParametersHelper;
         parent::__construct();
     }
 
@@ -43,15 +51,23 @@ class AnonymizeIpCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        if (!$this->coreParametersHelper->get('anonymize_ip')) {
+            return $this->exitWithError('Anonymization could not be done because anonymize Ip feature is disabled for this instance.', $output);
+        }
         try {
             $deletedRows = $this->ipAddressRepository->deleteAllIpAddress();
             $output->writeln(sprintf('<info>%s IP addresses have been deleted</info>', $deletedRows));
         } catch (DBALException $e) {
-            $output->writeln(sprintf('<error>Deletion of IP addresses failed because of database error: %s</error>', $e->getMessage()));
-
-            return 1;
+            return $this->exitWithError(sprintf('Deletion of IP addresses failed because of database error: %s', $e->getMessage()), $output);
         }
 
-        return 0;
+        return ExitCode::SUCCESS;
+    }
+
+    private function exitWithError(string $message, OutputInterface $output): int
+    {
+        $output->writeln(sprintf('<error>%s</error>', $message));
+
+        return ExitCode::FAILURE;
     }
 }
